사용자 회원가입 시 주소정보 등록 기능 추가

+++ src/main/java/com/festa/dto/EventDTO.java
@@ -11,7 +11,7 @@ import java.util.Date;
 @Builder
 public class EventDTO {
 
-    int eventNo;
+    long eventNo;
 
     long userNo;
